feat(mt5): finalize bridge integration coverage
Completes the bridge-first MT5 integration pass and captures the validation/docs updates needed to verify routing, deployment, and execution behavior from the main app.

+++ src/backtesting/output.py
@@ -23,7 +23,7 @@ class OutputBuilder:
         date_str: str,
         tickers: Sequence[str],
         agent_output: AgentOutput,
-        executed_trades: Mapping[str, int],
+        executed_trades: Mapping[str, float],
         current_prices: Mapping[str, float],
         portfolio: Portfolio,
         performance_metrics: Mapping[str, float | None],
